requirements.txt
attrs==21.2.0
Flask-Testing==0.8.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
toml==0.10.2
